Normalize subagent toolDefinitions to plain JSON Schema too
spawn-agent-inline builds the same state-stored toolDefinitions map as
loopAgentSteps and had the identical raw-inputSchema leak. Extract
toTokenCountInputSchema into util/to-json-schema.ts so both call sites
share one implementation (the util location avoids the import cycle
through run-agent-step, which re-exports for compatibility).
